+++ packages/xchain-thorchain-query/CHANGELOG.md
@@ -1,3 +1,15 @@
+# v0.6.10 (2023-11-07)
+
+## Fix
+
+- Trim memo for thorname registation
+
+# v0.6.9 (2023-11-05)
+
+## Update
+
+- Public validateAmount function
+
 # v0.6.8 (2023-11-04)
 
 ## Update